Add maximum number of session iterations
Adds support to the WebSurge GUI and CLI for the user to specify the
maximum number of times to repeat the requests in each session (per
thread).

This is useful for enabling scenarios where the user wants to generate a
precise payload, for example running each request in a session only once
using a single thread.